Move RTK handling to GPSManager
