+++ src/GPS/GPSManager.cc
@@ -24,6 +24,47 @@ GPSManager::GPSManager(QGCApplication* app, QGCToolbox* toolbox)
 GPSManager::~GPSManager()
 {
     disconnectGPS();
+
+    delete _gpsRtkFactGroup;
+    _gpsRtkFactGroup = nullptr;
+}
+
+void GPSManager::setToolbox(QGCToolbox *toolbox)
+{
+    QGCTool::setToolbox(toolbox);
+
+    _gpsRtkFactGroup = new GPSRTKFactGroup(this);
+
+    connect(this, &GPSManager::onConnect,          this, &GPSManager::_onGPSConnect);
+    connect(this, &GPSManager::onDisconnect,       this, &GPSManager::_onGPSDisconnect);
+    connect(this, &GPSManager::surveyInStatus,     this, &GPSManager::_gpsSurveyInStatus);
+    connect(this, &GPSManager::satelliteUpdate,    this, &GPSManager::_gpsNumSatellites);
+}
+
+void GPSManager::_onGPSConnect()
+{
+    _gpsRtkFactGroup->connected()->setRawValue(true);
+}
+
+void GPSManager::_onGPSDisconnect()
+{
+    _gpsRtkFactGroup->connected()->setRawValue(false);
+}
+
+void GPSManager::_gpsSurveyInStatus(float duration, float accuracyMM,  double latitude, double longitude, float altitude, bool valid, bool active)
+{
+    _gpsRtkFactGroup->currentDuration()->setRawValue(duration);
+    _gpsRtkFactGroup->currentAccuracy()->setRawValue(static_cast<double>(accuracyMM) / 1000.0);
+    _gpsRtkFactGroup->currentLatitude()->setRawValue(latitude);
+    _gpsRtkFactGroup->currentLongitude()->setRawValue(longitude);
+    _gpsRtkFactGroup->currentAltitude()->setRawValue(altitude);
+    _gpsRtkFactGroup->valid()->setRawValue(valid);
+    _gpsRtkFactGroup->active()->setRawValue(active);
+}
+
+void GPSManager::_gpsNumSatellites(int numSatellites)
+{
+    _gpsRtkFactGroup->numSatellites()->setRawValue(numSatellites);
 }
 
 void GPSManager::connectGPS(const QString& device, const QString& gps_type)
@@ -87,7 +128,6 @@ void GPSManager::disconnectGPS(void)
     _gpsProvider = nullptr;
     _rtcmMavlink = nullptr;
 }
-
 
 void GPSManager::GPSPositionUpdate(GPSPositionMessage msg)
 {
